+++ app/src/main/java/com/followapp/mytasks/adModule/model/AdRepository.kt
@@ -1,4 +1,23 @@
 package com.followapp.mytasks.adModule.model
 
+import com.google.android.gms.ads.AdRequest
+import com.google.android.gms.ads.AdView
+import com.google.android.gms.ads.AdListener
+import com.google.android.gms.ads.LoadAdError
+
 class AdRepository {
+
+    fun loadAd(adView: AdView, onAdLoaded: () -> Unit, onAdFailed: () -> Unit) {
+        val adRequest = AdRequest.Builder().build()
+        adView.loadAd(adRequest)
+        adView.adListener = object : AdListener() {
+            override fun onAdLoaded() {
+                onAdLoaded()
+            }
+
+            override fun onAdFailedToLoad(error: LoadAdError) {
+                onAdFailed()
+            }
+        }
+    }
 }
